[App][Domain][Infra] MapperController+Repositories+Roles

+++ src/main/java/com/fashionlike/proyecto_fashion_like/app/dto/UserDTO.java
@@ -6,9 +6,10 @@ import lombok.Data;
 @Data
 @Builder
 public class UserDTO {
-    private Long id;
+    private Integer id;
     private String name;
     private String username;
     private String password;
     private Boolean isActive;
+    //private String rol;
 }
